pyproject.toml
[project]
name = "abm-gods-eye"
version = "0.1.0"
description = "Drop-in LLM observer SDK for Python agent-based models"
readme = "README.md"
authors = [
    { name = "Gabriel Maayan", email = "gabemgem@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langgraph>=1.0.10",
    "mesa>=3.5.0",
    "networkx>=3.6.1",
]


[project.optional-dependencies]
mesa = ["mesa>=3.0"]
openai = [
    "langchain-openai>=0.3",
]
server = [
    "fastapi>=0.135.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
google = [
    "langchain-google-genai>=4.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
